factor out GrapheneMetadata from logs
