+++ python_modules/dagster-graphql/dagster_graphql/schema/dagster_types.py
@@ -3,7 +3,7 @@ from dagster import check
 from dagster.core.snap import PipelineSnapshot
 from dagster.core.types.dagster_type import DagsterTypeKind
 from dagster_graphql.implementation.events import iterate_metadata_entries
-from dagster_graphql.schema.logs.events import GrapheneMetadataEntry
+from dagster_graphql.schema.metadata import GrapheneMetadataEntry
 
 from .config_types import GrapheneConfigType, to_config_type
 from .errors import (
