+++ python_modules/dagster-graphql/dagster_graphql/schema/logs/__init__.py
@@ -3,19 +3,6 @@ def types():
     from .events import (
         GrapheneDisplayableEvent,
         GrapheneEngineEvent,
-        GrapheneTableSchemaMetadataEntry,
-        GrapheneTableMetadataEntry,
-        GrapheneFloatMetadataEntry,
-        GrapheneIntMetadataEntry,
-        GrapheneJsonMetadataEntry,
-        GrapheneMarkdownMetadataEntry,
-        GrapheneMetadataEntry,
-        GraphenePathMetadataEntry,
-        GraphenePythonArtifactMetadataEntry,
-        GrapheneTextMetadataEntry,
-        GrapheneUrlMetadataEntry,
-        GraphenePipelineRunMetadataEntry,
-        GrapheneAssetMetadataEntry,
         GrapheneExecutionStepFailureEvent,
         GrapheneExecutionStepInputEvent,
         GrapheneExecutionStepOutputEvent,
@@ -61,19 +48,6 @@ def types():
         GrapheneComputeLogs,
         GrapheneDisplayableEvent,
         GrapheneEngineEvent,
-        GraphenePipelineRunMetadataEntry,
-        GrapheneAssetMetadataEntry,
-        GrapheneTableSchemaMetadataEntry,
-        GrapheneTableMetadataEntry,
-        GrapheneFloatMetadataEntry,
-        GrapheneIntMetadataEntry,
-        GrapheneJsonMetadataEntry,
-        GrapheneMarkdownMetadataEntry,
-        GrapheneMetadataEntry,
-        GraphenePathMetadataEntry,
-        GraphenePythonArtifactMetadataEntry,
-        GrapheneTextMetadataEntry,
-        GrapheneUrlMetadataEntry,
         GrapheneExecutionStepFailureEvent,
         GrapheneExecutionStepInputEvent,
         GrapheneExecutionStepOutputEvent,
