+++ python_modules/dagster-graphql/dagster_graphql/schema/logs/events.py
@@ -9,8 +9,8 @@ from ...implementation.fetch_runs import get_run_by_id, get_step_stats
 from ...implementation.loader import BatchRunLoader
 from ..asset_key import GrapheneAssetKey, GrapheneAssetLineageInfo
 from ..errors import GraphenePythonError, GrapheneRunNotFoundError
+from ..metadata import GrapheneMetadataEntry
 from ..runs import GrapheneStepEventStatus
-from ..table import GrapheneTable, GrapheneTableSchema
 from ..util import non_null_list
 from .log_level import GrapheneLogLevel
 
@@ -26,14 +26,6 @@ class GrapheneMessageEvent(graphene.Interface):
 
     class Meta:
         name = "MessageEvent"
-
-
-class GrapheneMetadataEntry(graphene.Interface):
-    label = graphene.NonNull(graphene.String)
-    description = graphene.String()
-
-    class Meta:
-        name = "EventMetadataEntry"
 
 
 class GrapheneDisplayableEvent(graphene.Interface):
@@ -160,111 +152,6 @@ class GrapheneExecutionStepSkippedEvent(graphene.ObjectType):
     class Meta:
         interfaces = (GrapheneMessageEvent, GrapheneStepEvent)
         name = "ExecutionStepSkippedEvent"
-
-
-class GraphenePathMetadataEntry(graphene.ObjectType):
-    path = graphene.NonNull(graphene.String)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventPathMetadataEntry"
-
-
-class GrapheneTableMetadataEntry(graphene.ObjectType):
-    table = graphene.NonNull(GrapheneTable)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventTableMetadataEntry"
-
-
-class GrapheneTableSchemaMetadataEntry(graphene.ObjectType):
-
-    schema = graphene.NonNull(GrapheneTableSchema)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventTableSchemaMetadataEntry"
-
-
-class GrapheneJsonMetadataEntry(graphene.ObjectType):
-    jsonString = graphene.NonNull(graphene.String)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventJsonMetadataEntry"
-
-
-class GrapheneTextMetadataEntry(graphene.ObjectType):
-    text = graphene.NonNull(graphene.String)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventTextMetadataEntry"
-
-
-class GrapheneUrlMetadataEntry(graphene.ObjectType):
-    url = graphene.NonNull(graphene.String)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventUrlMetadataEntry"
-
-
-class GrapheneMarkdownMetadataEntry(graphene.ObjectType):
-    md_str = graphene.NonNull(graphene.String)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventMarkdownMetadataEntry"
-
-
-class GraphenePythonArtifactMetadataEntry(graphene.ObjectType):
-    module = graphene.NonNull(graphene.String)
-    name = graphene.NonNull(graphene.String)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventPythonArtifactMetadataEntry"
-
-
-class GrapheneFloatMetadataEntry(graphene.ObjectType):
-    floatValue = graphene.Field(graphene.Float)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventFloatMetadataEntry"
-
-
-class GrapheneIntMetadataEntry(graphene.ObjectType):
-    intValue = graphene.Field(
-        graphene.Int, description="Nullable to allow graceful degrade on > 32 bit numbers"
-    )
-    intRepr = graphene.NonNull(
-        graphene.String,
-        description="String representation of the int to support greater than 32 bit",
-    )
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventIntMetadataEntry"
-
-
-class GraphenePipelineRunMetadataEntry(graphene.ObjectType):
-    runId = graphene.NonNull(graphene.String)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventPipelineRunMetadataEntry"
-
-
-class GrapheneAssetMetadataEntry(graphene.ObjectType):
-    assetKey = graphene.NonNull(GrapheneAssetKey)
-
-    class Meta:
-        interfaces = (GrapheneMetadataEntry,)
-        name = "EventAssetMetadataEntry"
-
 
 class GrapheneObjectStoreOperationType(graphene.Enum):
     SET_OBJECT = "SET_OBJECT"
