+++ python_modules/dagster-graphql/dagster_graphql/schema/solids.py
@@ -8,7 +8,8 @@ from dagster.core.host_representation.historical import HistoricalPipeline
 from dagster.core.snap import CompositeSolidDefSnap, DependencyStructureIndex, SolidDefSnap
 from dagster.core.storage.pipeline_run import PipelineRunsFilter
 from dagster_graphql.implementation.events import iterate_metadata_entries
-from dagster_graphql.schema.logs.events import GrapheneMetadataEntry, GrapheneRunStepStats
+from dagster_graphql.schema.logs.events import GrapheneRunStepStats
+from dagster_graphql.schema.metadata import GrapheneMetadataEntry
 
 from .config_types import GrapheneConfigTypeField
 from .dagster_types import GrapheneDagsterType, to_dagster_type
